+++ classes/@classmodel/classmodel.m
@@ -1,15 +1,32 @@
 classdef classmodel < handle
-      
+   methods
+   end
+   
    methods(Static = true)
+      
+      function c = convertClasses(c, className)         
+         if isa(c, 'mdadata')
+            return
+         elseif islogical(c)
+            c = mdadata(c);         
+            if sum(c.valuesAll) == c.nRowsAll
+               c.factor(1, {className});
+            else
+               c.factor(1, {'None', className});
+            end
+         end   
+      end   
+      
       function newc = getClassFromFactor(c, className)
          factorLevelNames = c.factorLevelNames{1};
-         factorLevels = unique(c.valuesAll, 'stable');
+         factorLevels = unique(c.valuesAll);
                   
          if ischar(className)
             ind = ismember(factorLevelNames, className);
             if ~any(ind)
                error('Wrong class name!')
             end
+            
          elseif isnumeric(className)
             ind = className;
             if ind < 1 || ind > numel(factorLevels)
@@ -19,12 +36,17 @@ classdef classmodel < handle
          else
             error('Specify class name or number!')
          end   
-         
+
          newc = c.valuesAll == factorLevels(ind);
          newc = mdadata(newc, c.rowNamesAll, c.colNames, c.dimNames, c.name);
          newc.rowFullNames = c.rowFullNamesAll;
          newc.colFullNames = c.colFullNames;
-         newc.factor(1, {'None', className});
+         
+         if sum(newc.valuesAll) == c.nRowsAll
+            newc.factor(1, {className});
+         else
+            newc.factor(1, {'None', className});
+         end
          newc.excluderows(c.excludedRows);
       end   
    end